Refatoração para o padrão  PageObjects

+++ support/libs/mongo.rb
@@ -20,11 +20,11 @@ class MongoDB
     end
     # metodo para buscar o id do usuario 
 
-    def remove_equipo(name)
+    def remove_equipo(name, email)
         user_id = get_users(email)
         client = Mongo::Client.new('mongodb://rocklov-db:27017/rocklov')
         equipos = client[:equipos]
-        equipos.delete_many({name: name})
+        equipos.delete_many({name: name, user: user_id})
     end
     # metodo para remover o anuncio repitido antes de realizar um novo cadatsro ( teste)
